# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Heroes Empire gaming progression API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/add-xp:
    post:
      operationId: adminAddXp
      tags: [admin]
      summary: Grant XP to a player by Discord ID (bot)
      parameters:
        - in: header
          name: X-Bot-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAddXpBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/add-coins:
    post:
      operationId: adminAddCoins
      tags: [admin]
      summary: Grant coins to a player by Discord ID (bot)
      parameters:
        - in: header
          name: X-Bot-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAddCoinsBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/set-rockstar-id:
    post:
      operationId: adminSetRockstarId
      tags: [admin]
      summary: Set a player's Rockstar ID by Discord ID (bot)
      parameters:
        - in: header
          name: X-Bot-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSetRockstarIdBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /me/rockstar-id:
    patch:
      operationId: setRockstarId
      tags: [me]
      summary: Link or update the player's Rockstar Social Club ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRockstarIdBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /me/activity:
    post:
      operationId: logActivity
      tags: [activity]
      summary: Log a gameplay activity for the current player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogActivityBody"
      responses:
        "200":
          description: Activity result with rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResult"
...
  /me/equip:
    post:
      operationId: equipItem
      tags: [me]
      summary: Equip a cosmetic item from inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipItemBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /me/activities:
    get:
      operationId: listMyActivities
      tags: [me]
      summary: Recent activities for the current player
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
...
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a specific player profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
...
  /leaderboard/{type}:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get a leaderboard by category
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [xp, gta_missions, gta_heists, rl_matches, rl_wins]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Leaderboard rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leaderboard"
...
components:
  schemas:
...
    Attributes:
      type: object
      properties:
        unity:
          type: integer
        focus:
          type: integer
        strength:
          type: integer
        honor:
          type: integer
        passion:
          type: integer
        victory:
          type: integer
      required: [unity, focus, strength, honor, passion, victory]
    GameStats:
      type: object
      properties:
        gtaMissions:
          type: integer
        gtaHeists:
          type: integer
        rlMatches:
          type: integer
        rlWins:
          type: integer
      required: [gtaMissions, gtaHeists, rlMatches, rlWins]
    EquippedCosmetics:
      type: object
      properties:
        frameId:
          type: ["integer", "null"]
        badgeId:
          type: ["integer", "null"]
      required: [frameId, badgeId]
    Player:
      type: object
      properties:
        id:
          type: integer
        discordId:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        discordAvatar:
          type: ["string", "null"]
        discordDiscriminator:
          type: ["string", "null"]
        username:
          type: string
        handle:
          type: string
        avatarUrl:
          type: string
        rockstarId:
          type: ["string", "null"]
        totalXp:
          type: integer
        level:
          type: integer
        coins:
          type: integer
        rankTier:
          type: string
          enum: [ROOKIE, BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, MYTHIC]
        currentLevelXp:
          type: integer
          description: XP earned within the current level
        nextLevelXp:
          type: integer
          description: XP required to reach the next level (from current level start)
        levelProgressPct:
          type: number
        serverRank:
          type: integer
        gameStats:
          $ref: "#/components/schemas/GameStats"
        attributes:
          $ref: "#/components/schemas/Attributes"
        equipped:
          $ref: "#/components/schemas/EquippedCosmetics"
        streakDays:
          type: integer
        lastActiveAt:
          type: string
          format: date-time
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - discordId
        - discordUsername
        - discordAvatar
        - discordDiscriminator
        - username
        - handle
        - avatarUrl
        - rockstarId
        - totalXp
        - level
        - coins
        - rankTier
        - currentLevelXp
        - nextLevelXp
        - levelProgressPct
        - serverRank
        - gameStats
        - attributes
        - equipped
        - streakDays
        - lastActiveAt
        - joinedAt
    ActivityType:
      type: string
      enum: [gta_mission, gta_heist, rl_match, rl_win, daily_reward, level_up]
    LogActivityBody:
      type: object
      properties:
        type:
          type: string
          enum: [gta_mission, gta_heist, rl_match, rl_win]
      required: [type]
    ActivityResult:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"
        xpGained:
          type: integer
        coinsGained:
          type: integer
        leveledUp:
          type: boolean
        levelsGained:
          type: integer
        bonusCoins:
          type: integer
        activity:
          $ref: "#/components/schemas/ActivityEntry"
      required: [player, xpGained, coinsGained, leveledUp, levelsGained, bonusCoins, activity]
    ActivityEntry:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/ActivityType"
        xpGained:
          type: integer
        coinsGained:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, type, xpGained, coinsGained, createdAt]
...
    EquipItemBody:
      type: object
      properties:
        itemId:
          type: integer
      required: [itemId]
    SetRockstarIdBody:
      type: object
      properties:
        rockstarId:
          type: string
      required: [rockstarId]
...
    AdminAddXpBody:
      type: object
      properties:
        discordId:
          type: string
        amount:
          type: integer
      required: [discordId, amount]
    AdminAddCoinsBody:
      type: object
      properties:
        discordId:
          type: string
        amount:
          type: integer
      required: [discordId, amount]
    AdminSetRockstarIdBody:
      type: object
      properties:
        discordId:
          type: string
        rockstarId:
          type: string
      required: [discordId, rockstarId]
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        playerId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: string
        level:
          type: integer
        rankTier:
          type: string
          enum: [ROOKIE, BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, MYTHIC]
        value:
          type: integer
          description: The leaderboard's metric value (xp, missions, etc.)
        isMe:
          type: boolean
      required: [rank, playerId, username, avatarUrl, level, rankTier, value, isMe]
    Leaderboard:
      type: object
      properties:
        type:
          type: string
          enum: [xp, gta_missions, gta_heists, rl_matches, rl_wins]
        label:
          type: string
        valueLabel:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
      required: [type, label, valueLabel, entries]
...